#noticket Add Seamless Support for Current Code
