+++ src/main/java/hu/oe/nik/szfmv/automatedcar/virtualfunctionbus/VirtualFunctionBus.java
@@ -8,13 +8,10 @@ import hu.oe.nik.szfmv.automatedcar.virtualfunctionbus.packets.powertrain.ICarMo
 import hu.oe.nik.szfmv.automatedcar.virtualfunctionbus.packets.powertrain.IEngineStatusPacket;
 import hu.oe.nik.szfmv.automatedcar.virtualfunctionbus.packets.visualization.*;
 
-/**
- * This is the interface for the Virtual Function Bus. Components are only
- * allowed to collect sensory data exclusively using the VFB. The VFB stores the
- * input and output signals, inputs only have setters, while outputs only have
- * getters respectively.
- */
-public abstract class VirtualFunctionBus {
+import java.util.ArrayList;
+import java.util.List;
+
+public class VirtualFunctionBus {
 
     public ToPowerTrainPacket toPowerTrainPacket = new ToPowerTrainPacket();
     public GuiInputPacket guiInputPacket = new GuiInputPacket();
@@ -35,16 +32,30 @@ public abstract class VirtualFunctionBus {
     public IParkingRadarDisplayStatePacket parkingRadarDisplayStatePacket;
     public ISelectedDebugListPacket selectedDebugListPacket;
 
-    public abstract void registerComponent(SystemComponent component);
+    /**All registered components.*/
+    private final List<SystemComponent> components = new ArrayList<>();
 
-    public abstract void loop();
+    /**
+     * Registers the provided {@link SystemComponent component}.
+     *
+     * @param component component to be registed.
+     */
+    public void registerComponent(SystemComponent component) {
+        if (component.getClass().getAnnotation(DependsOn.class) != null) {
+            throw new IllegalArgumentException("System component '" + component.getClass().getSimpleName()
+                    + "' has dependencies, please use DependentVirtualFunctionBus for it instead!");
+        }
 
-    public static SimpleVirtualFunctionBus createNewSimpleVFB() {
-        return new SimpleVirtualFunctionBus();
+        components.add(component);
     }
 
-    public static DependentVirtualFunctionBus createNewVFBSupportingDependencies() {
-        return new DependentVirtualFunctionBus();
+    /**
+     * Calls cyclically the registered {@link SystemComponent component}s once the virtual function bus has started.
+     */
+    public void loop() {
+        for (SystemComponent component : components) {
+            component.loop();
+        }
     }
 
 }
